refactor(terminal): extract ZoomBadge / BroadcastBadge into StatusBadges.swift
Both badges were file-private NSView subclasses pinned to the bottom of
TerminalContainerView. They have no host coupling beyond Theme reads, so they
move to their own file. Made internal so the host can construct them by name.
